Graphics Stable
Switched delay to a swing.Timer - Graphics no longer crash when the previously mentioned exception is thrown. The exception doesn't seem to have any negative consequences, so I'm not worried about it right now.

Switched all Graphics calls to Graphics2D

Added code (currently commented out) for moving the JFrame to a secondary screen

Added lines to the thermometer for a visual cue as to how much has been collected/pledged

Added Charity's logo

+++ src/main/java/Thermometer.java
@@ -69,7 +69,7 @@ public class Thermometer {
         // Above code utilized from Google Sheets API Quickstart file
         // Below code modified from above file for thermometer use
 
-        int totalRaised = 0;
+        double totalRaised = 0;
         int rowsTraversed = 0;
         int i;
 
@@ -87,7 +87,7 @@ public class Thermometer {
                 if(i > rowsTraversed && !(row.get(0).equals("EXIT"))) { // read row if row previously unread and if column 0 does not contain "EXIT"
                     rowsTraversed++; //track rows read
                     window.addDonor((String) row.get(0), Double.valueOf((String) row.get(1)));
-                    totalRaised += Integer.valueOf((String) row.get(1)); // track total raised for use with graphical thermometer
+                    totalRaised += Double.valueOf((String) row.get(1)); // track total raised for use with graphical thermometer
                     System.out.println(row.get(0) + " donated $" + row.get(1) + " || total = " + totalRaised); // temporary back-end debug line
                     //update graphics
                 }else if(row.get(0).equals("EXIT")) break; // break for loop if column 0 contains "EXIT" (To deny access to rows that may not be completed yet)
